third_party/blink/web_tests/virtual/select-parser-relaxation-opt-out/external/wpt/html/semantics/forms/the-select-element/customizable-select/switch-picker-appearance.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] Basic functionality of select picker and appearance
  assert_equals: expected "rgb(255, 0, 0)" but got "rgba(0, 0, 0, 0)"
[FAIL] Basic functionality of select picker with appearance:auto
  assert_equals: expected "auto" but got "none"
[FAIL] Basic functionality of select picker with appearance:none
  assert_equals: appearance:none picker is never open expected "rgb(255, 0, 0)" but got "rgba(0, 0, 0, 0)"
[FAIL] Switching appearance in popover-open should close the picker
  assert_equals: expected "base-select" but got "none"
[FAIL] Switching appearance in JS after picker is open should close the picker
  assert_equals: expected "base-select" but got "none"
[FAIL] Test of the test harness
  assert_equals: expected "base-select" but got "auto"
[FAIL] The select picker is closed if the <select> appearance value is changed via CSS while the picker is open
  assert_equals: expected "base-select" but got "auto"
[FAIL] The select picker is closed if the ::picker() appearance value is changed via CSS while the picker is open
  assert_equals: expected "base-select" but got "auto"
[FAIL] The select picker is closed if the <select> inline appearance value is changed while the picker is open
  assert_equals: expected "base-select" but got "auto"
Harness: the test ran to completion.
